requirements.txt
Flask==2.2.3
requests==2.28.2
pygame==2.5.1
gunicorn==21.2.0
python-dotenv==1.0.0
